[Clang][AMDGPU] Emit stub version of OpenCL Kernel
OpenCL allows a kernel function to call another kernel function.
To facilitate this we emit a stub version of each kernel function
with different name mangling scheme, and replace the kernel
callsite appropriately.

This commit fixes https://github.com/llvm/llvm-project/issues/60313
https://ontrack-internal.amd.com/browse/SWDEV-245936

D120566 was an earlier effort to upstream a fix for this issue.